Move get_timeline_at_date to new mica.utils
